Add files via upload
Ps2a is the first half of ps2, the implementation of the solar system. A text file including the coordinates of the first five planets in our solar system are passed into the program and a window will display their images in accordance. ./NBody < planets.txt is the stdin command used on the command line in order to cin into the stream and retrieve the radius and window dimension.