Add files via upload
The rest of ps2b involves adding the function step() to imitate the planets rotating around the sun. As a line of stdin is given, ./NBody 157788000.0 25000.0 < planets.txt will read into the program, open a window to display the solar system, and once it has stopped, the coordinates of each celestialbody will be printed in the terminal.